Prior to fixing texture bug
