+++ sdk/engine/source/graphics/meshes/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(${NC_ENGINE_LIB}
+    PRIVATE
+        MeshStorage.cpp
+)
