+++ sdk/engine/source/graphics/resources/ShaderResourceService.h
@@ -13,7 +13,7 @@ namespace nc
         struct ShadowMap;
         class Graphics;
         struct ObjectData;
-        struct TextureData;
+        struct TextureImageInfo;
         class CubeMap;
     }
 }
