+++ sdk/engine/source/graphics/textures/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${NC_ENGINE_LIB}
+    PRIVATE
+        TextureManager.cpp
+        TextureStorage.cpp
+)
