+++ sdk/sample/config.ini
@@ -30,4 +30,4 @@ near_clip=0.5
 far_clip=400
 use_shadows=1
 antialiasing=8
-use_validation_layers=0
+use_validation_layers=1
